+++ src/types/Configs.ts
